rename dict.py
