+++ momaapi/data/__init__.py
@@ -1,2 +1,2 @@
 from .ann import Metadatum, Act, SAct, HOI, Clip, BBox, Entity, Predicate
-from .dict import Bidict, OrderedBidict, LazyDict
+from .dicts import Bidict, OrderedBidict, LazyDict
